working-poor-silicon-valley/requirements.txt
# Working Poor Silicon Valley Analysis
# Python 3.9+

# Geospatial
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# API access
requests>=2.31.0
python-dotenv>=1.0.0

# Census data
census>=0.8.21

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars (optional)
tqdm>=4.66.0
